requirements.txt
astor==0.8.1
cycler==0.11.0
fonttools==4.29.1
kiwisolver==1.3.2
matplotlib==3.5.1
networkx==2.6.3
numpy==1.22.2
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
PyYAML==6.0
scipy==1.8.0
six==1.16.0
tqdm==4.62.3
pyDOE2~=1.3.0